GSS dataset for in-class exercise

The General Social Survey polls individuals about their attitudes and beliefs. This dataset is contains a selection of variables.

Usage

gss

Format

A data frame with 2,348 rows and 27 columns:

id

Respondent ID number.

wtss

Weight variable.

age

Age of respondent.

backpain

In the past 12 months, has the respondent had back pain every day for a week or more?

colath

Should a person who is against all churches and religion be allowed to teach in a college or university, or not?

colmslm

Should a Muslim clergyman who preaches hatred of the United States be allowed to teach in a college or university, or not?

courts

In general, do you think the courts in this area deal too harshly or not harshly enough with criminals?

degree

Highest degree of respondent.

grass

Do you think the use of marijuana should be made legal or not?

gunlaw

Would you favor or oppose a law which would require a person to obtain a police permit before he or she could buy a gun?

happy

Taken all together, how would you say things are these days -- would you say that you are very happy, pretty happy, or not too happy?

hispanic

Are you Spanish, Hispanic, or Latino/Latina? If yes, which group are you from?

hrs1

How many hours did you work last week, at all jobs?

income16

Total family income.

letdie1

When a person has a disease that cannot be cured, do you think Doctors should be allowed by law to end the patient's life by some painless means if the patient and his family request it?

owngun

Do you happen to have in your home any guns or revolvers?

partyid

Generally speaking, do you usually think of yourself as a Republican, Democrat, Independent, or what?

polviews

I'm going to show you a seven-point scale on which the political views that people might hold are arranged from extremely liberal -- point 1 -- to extremely conservative -- point 7. Where would you place yourself on this scale?

pray

About how often do you pray?

pres16

Did you vote for Clinton or Trump?

race

What race do you consider yourself?

region

Region of interview.

sex

Respondent's sex.

sexfreq

About how often did you have sex during the last 12 months?

wordsum

Total number of correct words on a vocabulary test.

zodiac

Astrological sign of respondent.

Source

2018 General Social Survey.